+++ src/Craned/TaskManager.h
@@ -375,9 +375,9 @@ class TaskManager {
   // and doesn't have the ownership of underlying objects.
   // A TaskInstance may contain more than one ProcessInstance.
   absl::flat_hash_map<uint32_t /*pid*/, TaskInstance*> m_pid_task_map_
-      GUARDED_BY(m_mtx_);
+      ABSL_GUARDED_BY(m_mtx_);
   absl::flat_hash_map<uint32_t /*pid*/, ProcessInstance*> m_pid_proc_map_
-      GUARDED_BY(m_mtx_);
+      ABSL_GUARDED_BY(m_mtx_);
 
   absl::Mutex m_mtx_;
